% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bor.R
\name{bor_nm2chr}
\alias{bor_nm2chr}
\title{Extract the shorthand lake name from a string}
\usage{
bor_nm2chr(x)
}
\arguments{
\item{x}{\code{chr}}
}
\value{
\code{chr}
}
\description{
Extract the shorthand lake name from a string
}
\examples{
bor_nm2chr("Mead Annual Pool Elevation")
}
