% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia_crb_dam_data.R
\name{crb_dam_reservoir}
\alias{crb_dam_reservoir}
\title{Look up reservoir or dam}
\usage{
crb_dam_reservoir(dam = NULL, reservoir = NULL)
}
\arguments{
\item{dam}{\code{(chr)} Name of dam}

\item{reservoir}{\code{(chr)} Name of reservoir}
}
\value{
\code{(tbl)}
}
\description{
If no dam or reservoir supplied, the full table of options is returned
}
\examples{
crb_dam_reservoir("Blue Mesa")
}
