% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_data.R
\name{dam_fetch}
\alias{dam_fetch}
\title{Fetches dam data from BOR website}
\usage{
dam_fetch(path)
}
\arguments{
\item{path}{\code{(chr)} path to dam info page}
}
\value{
General, Dimensions, Hydraulics & Hydrology tables
}
\description{
Fetches dam data from BOR website
}
