% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_bor_schema.R
\docType{data}
\name{dam_groupings}
\alias{dam_groupings}
\title{Groupings of Dams in the CRB}
\format{
An object of class \code{list} of length 3.
}
\usage{
dam_groupings
}
\description{
These groupings are derived from this \href{https://www.usbr.gov/uc/rm/crsp/index.html}{link}
}
\keyword{datasets}
