% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_data.R
\name{dam_info}
\alias{dam_info}
\title{Get details on a dam}
\usage{
dam_info(dam)
}
\arguments{
\item{dam}{\code{(chr)} Name of the dam to lookup. If not supplied, all tables will be returned}
}
\value{
\code{(list)} of tables
\itemize{
\item{\code{General}}
\item{\code{Dimensions}}
\item{\code{Hydraulics & Hydrology}}
}
}
\description{
See \code{dam_select} for available dams
}
\examples{
dam_tables("Blue Mesa")
}
