% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_data.R
\name{dam_select}
\alias{dam_select}
\title{Fuzzy search for a dam name or view all}
\usage{
dam_select(name)
}
\arguments{
\item{name}{\code{(chr)} of the dam. \strong{Optional} leave blank to return all options}
}
\value{
\code{(named chr)} with the url named according to the dam
}
\description{
Fuzzy search for a dam name or view all
}
