% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_data.R
\name{dam_table_parse}
\alias{dam_table_parse}
\title{Parse downloaded dam data rds files}
\usage{
dam_table_parse()
}
\value{
Generates \emph{R/all_dams.R}
}
\description{
Parse downloaded dam data rds files
}
