% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_data.R
\name{dam_table_update}
\alias{dam_table_update}
\title{Update info rds for all dams}
\usage{
dam_table_update(dams = dam_select())
}
\arguments{
\item{dams}{\code{(chr)} named character of url paths}
}
\value{
All saved rds files in \emph{data/dams} and an updated \emph{R/all_dams.R}
}
\description{
Update info rds for all dams
}
