% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dam_data.R
\name{dam_tables}
\alias{dam_tables}
\title{Get details on a dam}
\usage{
dam_tables(dam)
}
\arguments{
\item{dam}{\code{(chr)} Name of the dam to lookup}
}
\value{
\code{(list)} of tables
\itemize{
\item{\code{General}}
\item{\code{Dimensions}}
\item{\code{Hydraulics & Hydrology}}
}
}
\description{
See \code{dam_options} for available dams
}
\examples{
dam_tables("Blue Mesa")
}
