% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{dtable}
\alias{dtable}
\title{dtable}
\usage{
dtable(x, bset, bsep = ".", asep = ";", missing = "x", noc = "0")
}
\arguments{
\item{x}{Behavior observation field record. An object with one row per observation and 3 columns: id of the subject under observation, the action observed and the observer id. Action entries may refer to the interaction partner and the behavior observed: [behavior][bsep][partner id] if partner is the target or [partner id][bsep][behavior] if subject is the target.}

\item{bset}{A vector specifying the set of behaviors observed.}

\item{bsep}{A separation character for bset elements and id's. Default is ".".}

\item{asep}{A separation character for multiple actions in just one entry. Default is ";".}

\item{missing}{A character specifying missing subjects. Default is "x".}

\item{noc}{A character specifying no ocorrences. Default is "0".}
}
\value{
dtable returns a data frame with 7 columns: "id1" indicates the subject under observation and "id2" it's interaction partner. "sender_id1" takes the values 1 or 0 whether id1 is the sender or not. If not, id2 is the sender. "behavior" indicates the behavior observed. If no ocurrences or misses where observed, "no_ocurrence" and "missing" take the value 1. All entries of multiple actions are separated in consecutive rows. All characters are changed to lower case and any white spaces are removed.
}
\description{
dtable takes an object representing the data record of field behavior observations of dyadic social interactions and returns a data frame of the same data organized in a way suitable for further statistical analysis.
}
\examples{

## converting the field raw data of behavior observations to suitable form useful for further statistical analysis
bset <- c("+","-")
data <- dtable(ex_field_data, bset)
}
