% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res_hydrodata.R
\name{get_res_data}
\alias{get_res_data}
\title{Get Reservoir Hydrodata from BoR Pseudo-API}
\usage{
get_res_data(reservoir_num, data_type_num, date = NULL, most_recent = FALSE)
}
\arguments{
\item{reservoir_num}{\code{(num)} numeric value corresponding to a given reservoir}

\item{data_type_num}{\code{(num)} numeric value correspond to a given data type}

\item{date}{\code{(chr)} date to filter for in YYYY-MM-DD format}
}
\value{
a two-column tibble of returned data
}
\description{
Grab any of the reservoir data from the BoR Hydrodata API if you know the
number corresponding to a given reservoir and the number corresponding to a
given datatype. The function also allows you to grab a specific day, or the
entire historic/current data if no date is specified.
}
\examples{
get_res_data(919, 49) # get Powell Elevation
}
