% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_bor_schema.R
\docType{data}
\name{mead_powell_slider_defaults}
\alias{mead_powell_slider_defaults}
\title{Default Elevations for Mead/Powell for use in sliders}
\format{
An object of class \code{data.frame} with 2 rows and 5 columns.
}
\usage{
mead_powell_slider_defaults
}
\description{
Default Elevations for Mead/Powell for use in sliders
}
\keyword{datasets}
