% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia_crb_dam_data.R
\name{update_crb_dams}
\alias{update_crb_dams}
\title{Update the local version of the tables found at \href{list of dams in the colorado river system}{https://en.wikipedia.org/wiki/List_of_dams_in_the_Colorado_River_system}}
\usage{
update_crb_dams(
  .url = "https://en.wikipedia.org/wiki/List_of_dams_in_the_Colorado_River_system"
)
}
\arguments{
\item{.url}{\code{(chr)} The wikipedia url}
}
\description{
Update the local version of the tables found at \href{list of dams in the colorado river system}{https://en.wikipedia.org/wiki/List_of_dams_in_the_Colorado_River_system}
}
