% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_bor.R
\name{update_powell_mead_elevations}
\alias{update_powell_mead_elevations}
\title{Update \code{powell_mead_elevations} table}
\usage{
update_powell_mead_elevations()
}
\value{
\code{file}
}
\description{
A table of key elevations pertaining to Lake's Mead & Powell
}
