% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countb.R
\name{countb}
\alias{countb}
\title{countb}
\usage{
countb(x)
}
\arguments{
\item{x}{A data frame with 7 columns: the 1st column indicating the subject under observation and the 2nd the interaction partner (if a partner exists, otherwise NA), the 3rd indicating the sender (with 1 refering to id1 and 0 to id2), the 4th the behavior observed, the 5th no ocurrences (taking the value 1 or NA otherwise), the 6th missing subjects (taking the value 1 or NA otherwise) and 7th the observer id.}
}
\value{
countb returns a list of observers, each with a list of behaviors, each with a data frame with interaction countings for each dyad, with row subjects indicating senders and column subjects indicating receivers. All characters are changed to lower case and any white spaces are removed.
}
\description{
countb takes a data frame of field behavior observations of dyadic social interactions and returns interaction countings for each dyad x behavior x observer.
}
\examples{

 ## Considering ex_field_data, count the number of positive and negative interactions between the various subjects S1, s2, ..., s9 observed by observers 1 and 2. But first converting the raw data to suitable form of input for countb()
 bset <- c("+","-")
 input <- dtable(ex_field_data, bset)
 observations <- countb(input)
}
